Move core Embedly classes to rover package
