+++ kifi-backend/modules/scraper/test/com/keepit/scraper/embedly/FakeEmbedlyClient.scala
@@ -3,6 +3,5 @@ package com.keepit.scraper.embedly
 import scala.concurrent.Future
 
 class FakeEmbedlyClient extends EmbedlyClient {
-  override def embedlyUrl(url: String): String = "https://www.google.com"
   override def getEmbedlyInfo(url: String): Future[Option[EmbedlyInfo]] = Future.successful(None)
 }
